Relocate proptype decls in frontend; add no prior experience check
